third_party/blink/web_tests/accessibility/color-changed.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/run-after-layout-and-paint.js"></script>

<p id="blue" style="color: blue">Blue foreground.</p>

<script>
const kBlueColor = 0xFF;
const kGreenColor = 0x8000;

  test_after_layout_and_paint(() => {
    let blue = document.getElementById('blue');
    let axBlue = accessibilityController.accessibleElementById('blue');
    assert_equals(axBlue.color, kBlueColor);
    blue.style.color = 'green';
    assert_equals(axBlue.color, kGreenColor);
  }, 'Ensures that the color can be dynamically changed.');
</script>
